pyproject.toml
[tool.poetry]
name = "restful-vdi"
version = "0.1.0"
description = ""
authors = ["2604801 <brent.hamilton@cvshealth.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.3.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.0"
pytest = "^8.2.2"
gunicorn = "^22.0.0"
pymssql = "^2.3.0"
pyodbc = "^5.1.0"
fastapi-sso = "^0.15.0"
msal = "^1.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
